requirements.txt
numpy
Pillow
facenet-pytorch
Flask
flask-restplus
joblib
seaborn
scikit-learn
torch
torchvision
Werkzeug
opencv-python
matplotlib
streamlit
huggingface-hub
google-api-python-client
oauth2client
streamlit-webrtc